2_document_summarizer/requirements.txt
python-dotenv
langchain
ipykernel
pypdf
langchain_community
langchain-community
bs4
langchain-text-splitters
wikipedia
PyMuPDF
langchain-openai
chromadb
faiss-cpu
streamlit
langchain_groq
langchain_core
fastapi
langserve
langchain_community
langchain_huggingface
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
sentence-transformers
faiss-cpu
